add rotation for joints
Revolute joints does not seem to have an api for
setting the local frame which prevents rotation of the joint.
A way around this without change the rapier source is to set the
local frames on the GenericJoint directly.

Also added a car model to test the joints and to highlight flaws
in the spawning api.

also increased the default friciton and set default restitution to zero