add global orientation and angular velocity
The orientation and anggular velocity is now stored in the multiroot component.
It seems that is things that keeping everything in its own component will result in
huge queries. I'm not sure I like that. Is there any benifit? For the first iteration of I will
keep everything connect to multibodies in the root and child component