Initial python api prototype (#103)
A python frontend communicates with Kesko over a http using a client server architecture. There is a new kesko_tcp crate that adds a tcp server in Kesko and only steps forward when requested to do so by the client. Perhaps in the future the responsibilities should be changed and the python front end is the server and Kesko the client in order to be able to easy spawn multiple Kesko instances and connect to one training run.

In this initial commit it is possible to step the simulation, pause/run the physics, spawn models and get back the states for each model.