tcp communication stages
In order to anwser a tcp request on one cycle the request and response system are devided
into to systems that are in separate stages. The request stage is run before anything else
and the response system is the last system that runs in order to assure that all data is
available and updated. Much polish to do however.

Also the requests goes through SystemEvents which is meant to be the API
into the simulator. Will see if this is a good idea or not but for now
it makes things a bit cleaner.