acceleration based camera movements (#119)
all movement are now acceleration based giving them a mush smoother feel. all accelerations can be set for the individual movements. The movements are still event based so that in the future camera movement such as object tracking an fly-to mechanics can be easily implemented.

new is also that translation can be controlled by the arrow keys.