Joint rewrite (#121)
joints are re-written to use the build pattern in a similar way as rapier, but stores and updates useful joint specific variables. This removes the base Joint class which never should have existed in the first place.

Also moved to fork of rapier with a hot fix for building multibodies. Will investigate this further.