change erp for the physics (#128)
the high constraint violation setting of 0.99 for contact penetration caused instabilities when using f32. setting it to 0.8 seems to fix this.

also setting mass for colliders seemed to cause a strange behavior when it came to gravity for multibodies. instead setting additional mass for the rigidbody made gravity behave as expected.