[python api] collision events (#130)
Collision events are now sent to the python frontend. The collision events in kesko are serialized so no new data structure is added.

Also this pr includes a lot of refatctoring of the python code. Mainly adding de-serialization of responses in order to make environments easier to create.

Note: python api is still in prototype state.

closes: #124