PyKesko packaging and build action (#168)
Restructuring the project layout to have kesko and pykesko in separate folders, this in order to make it less confusing and also be able to have python code and rust code in the same folder.

This also add a build action for PyKesko that build wheels for Linux, Mac and Windows. See build_kesko.toml for more info. Currently it does not upload to pypi and can only be manually triggered. This will change when the final release workflow is in place.